packages/faasit-examples/examples/dnn-train/code/dpflow/requirements.txt
pydantic>=1.10
cachetools>=5.3.0
python-snappy>=0.7.1